requirements.txt
pydantic-settings
pyppeteer
textual-dev
nest_asyncio